Moved tuner react files to their own folder
